+++ front/src/components/calendar/year/MonthCalendar/CalendarBody/Date/index.jsx
@@ -8,7 +8,7 @@ import Moment from '../../../../../../utils/moment';
 import styles from './style.module.css';
 import { getAllCalendarAndEvent } from '../../../../../../store/thunk';
 import { eventsByDateSelector } from '../../../../../../store/selectors/events';
-import { triggerDOM } from '../../../../../modal/EventListModal';
+import { triggerDOM } from '../../../../../../modal/component/EventListModal';
 import { EventListModalContext } from '../../../../../../context/EventModalContext';
 
 const Index = ({ month, date }) => {
