+++ front/src/modal/layout/CreateModalLayout/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { CreateEventModalContext } from '../../../../context/EventModalContext';
-import CreateEventModal from '../../CreateEventModal';
-import useEventModal from '../../../../hooks/useEventModal';
+import CreateEventModal from '../../component/CreateEventModal';
+import { CreateEventModalContext } from '../../../context/EventModalContext';
+import useEventModal from '../../../hooks/useEventModal';
 
 const Index = ({ children }) => {
   const { isModalShown, modalData, setModalData, showModal, hideModal } =
